jd_assistants/pyproject.toml
[project]
name = "jd_assistants"
version = "0.1.0"
description = "HR Assistant using Langchain and Langgraph"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-groq",
    "langchain-community",
    "pymupdf",
    "pdfplumber",
    "pandas",
    "python-dotenv",
    "termcolor",
    "fastapi",
    "uvicorn[standard]",
    "gradio",
    "sqlalchemy",
    "asyncpg",
    "redis",
    "plotly",
    "python-multipart",
    "alembic",
    "aiosqlite"
]



[project.scripts]
jd_assistants = "jd_assistants.main:run"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jd_assistants"]
